refactor: reorg generators packages and simplify namings
