+++ pkg/generator/appconfiguration/generator/namespace_generator.go
@@ -1,4 +1,4 @@
-package generators
+package generator
 
 import (
 	"fmt"
@@ -6,6 +6,7 @@ import (
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
+	"kusionstack.io/kusion/pkg/generator/appconfiguration"
 	"kusionstack.io/kusion/pkg/models"
 )
 
@@ -13,7 +14,7 @@ type namespaceGenerator struct {
 	projectName string
 }
 
-func NewNamespaceGenerator(projectName string) (Generator, error) {
+func NewNamespaceGenerator(projectName string) (appconfiguration.Generator, error) {
 	if len(projectName) == 0 {
 		return nil, fmt.Errorf("project name must not be empty")
 	}
@@ -23,8 +24,8 @@ func NewNamespaceGenerator(projectName string) (Generator, error) {
 	}, nil
 }
 
-func NewNamespaceGeneratorFunc(projectName string) NewGeneratorFunc {
-	return func() (Generator, error) {
+func NewNamespaceGeneratorFunc(projectName string) appconfiguration.NewGeneratorFunc {
+	return func() (appconfiguration.Generator, error) {
 		return NewNamespaceGenerator(projectName)
 	}
 }
@@ -42,8 +43,8 @@ func (g *namespaceGenerator) Generate(spec *models.Spec) error {
 		ObjectMeta: metav1.ObjectMeta{Name: g.projectName},
 	}
 
-	return appendToSpec(
-		kubernetesResourceID(ns.TypeMeta, ns.ObjectMeta),
+	return appconfiguration.AppendToSpec(
+		appconfiguration.KubernetesResourceID(ns.TypeMeta, ns.ObjectMeta),
 		ns,
 		spec,
 	)
