+++ pkg/generator/appconfiguration/generator/namespace_generator_test.go
@@ -1,4 +1,4 @@
-package generators
+package generator
 
 import (
 	"testing"
